requirements.txt
torch>=2.0.0
torchvision>=0.15.0
diffusers>=0.20.0
transformers>=4.30.0
accelerate>=0.20.0
PyYAML
matplotlib
numpy
tqdm
scipy
ftfy
regex
